Created currentBookingsController.ts
